+++ noodles-cram/src/record/mate_flags.rs
@@ -1,7 +1,7 @@
 bitflags::bitflags! {
     /// CRAM record next mate flags.
     #[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-    pub struct NextMateFlags: u8 {
+    pub struct MateFlags: u8 {
         /// The mate is on the negative strand (`0x01`).
         const ON_NEGATIVE_STRAND = 0x01;
         /// The mate is unmapped (`0x02`).
@@ -9,14 +9,14 @@ bitflags::bitflags! {
     }
 }
 
-impl NextMateFlags {
+impl MateFlags {
     /// Returns whether the `ON_NEGATIVE_STRAND` flag is set.
     ///
     /// # Examples
     ///
     /// ```
-    /// use noodles_cram::record::NextMateFlags;
-    /// assert!(NextMateFlags::ON_NEGATIVE_STRAND.is_on_negative_strand());
+    /// use noodles_cram::record::MateFlags;
+    /// assert!(MateFlags::ON_NEGATIVE_STRAND.is_on_negative_strand());
     /// ```
     pub fn is_on_negative_strand(self) -> bool {
         self.contains(Self::ON_NEGATIVE_STRAND)
@@ -27,22 +27,22 @@ impl NextMateFlags {
     /// # Examples
     ///
     /// ```
-    /// use noodles_cram::record::NextMateFlags;
-    /// assert!(NextMateFlags::UNMAPPED.is_unmapped());
+    /// use noodles_cram::record::MateFlags;
+    /// assert!(MateFlags::UNMAPPED.is_unmapped());
     /// ```
     pub fn is_unmapped(self) -> bool {
         self.contains(Self::UNMAPPED)
     }
 }
 
-impl From<u8> for NextMateFlags {
+impl From<u8> for MateFlags {
     fn from(value: u8) -> Self {
         Self::from_bits_truncate(value)
     }
 }
 
-impl From<NextMateFlags> for u8 {
-    fn from(flags: NextMateFlags) -> Self {
+impl From<MateFlags> for u8 {
+    fn from(flags: MateFlags) -> Self {
         flags.bits()
     }
 }
@@ -53,7 +53,7 @@ mod tests {
 
     #[test]
     fn test_default() {
-        let flags = NextMateFlags::default();
+        let flags = MateFlags::default();
 
         assert!(flags.is_empty());
 
@@ -63,19 +63,19 @@ mod tests {
 
     #[test]
     fn test_contains() {
-        assert!(NextMateFlags::ON_NEGATIVE_STRAND.is_on_negative_strand());
-        assert!(NextMateFlags::UNMAPPED.is_unmapped());
+        assert!(MateFlags::ON_NEGATIVE_STRAND.is_on_negative_strand());
+        assert!(MateFlags::UNMAPPED.is_unmapped());
     }
 
     #[test]
     fn test_from_u8_for_flags() {
-        assert_eq!(NextMateFlags::from(0x01), NextMateFlags::ON_NEGATIVE_STRAND);
-        assert_eq!(NextMateFlags::from(0x02), NextMateFlags::UNMAPPED);
+        assert_eq!(MateFlags::from(0x01), MateFlags::ON_NEGATIVE_STRAND);
+        assert_eq!(MateFlags::from(0x02), MateFlags::UNMAPPED);
     }
 
     #[test]
     fn test_from_flags_for_u8() {
-        assert_eq!(u8::from(NextMateFlags::from(0x01)), 0x01);
-        assert_eq!(u8::from(NextMateFlags::from(0x02)), 0x02);
+        assert_eq!(u8::from(MateFlags::from(0x01)), 0x01);
+        assert_eq!(u8::from(MateFlags::from(0x02)), 0x02);
     }
 }